+++ pkg/testsuite/handler_querydata_test.go
@@ -112,7 +112,7 @@ func TestAuthentication(t *testing.T) {
 			metaData := res.Frames[0].Meta.Custom.(*infinity.CustomMeta)
 			require.NotNil(t, res.Error)
 			require.NotNil(t, metaData)
-			require.Equal(t, "unsuccessful HTTP response. 401 Unauthorized", metaData.Error)
+			require.Equal(t, "unsuccessful HTTP response code\nstatus code : 401 Unauthorized", metaData.Error)
 			require.Equal(t, http.StatusUnauthorized, metaData.ResponseCodeFromServer)
 		})
 	})
@@ -707,8 +707,8 @@ func TestInlineSources(t *testing.T) {
 			skipGoldenCheck: true,
 			test: func(t *testing.T, frame *data.Frame) {
 				require.NotNil(t, frame)
-				assert.Equal(t, data.NewField("Sex", nil, []*string{toSP("Male"), toSP(""), toSP("Others")}), frame.Fields[0])
-				assert.Equal(t, data.NewField("Count by gender", nil, []*float64{toFP(2), toFP(1), toFP(1)}), frame.Fields[1])
+				assert.Equal(t, data.NewField("Sex", nil, []*string{toSP("Male"), toSP("Others")}), frame.Fields[0])
+				assert.Equal(t, data.NewField("Count by gender", nil, []*float64{toFP(2), toFP(1)}), frame.Fields[1])
 			},
 		},
 		{